+++ src/core/computed/get-define-computed-handler.js
@@ -1,19 +1,14 @@
 import configureDepFns from '../base/configure-dep-fns';
-import { CATE_REF } from '../../support/constant';
+import { CATE_REF, FN_CU } from '../../support/constant';
 
-let sortFactor = 1;
 
-export default function (refCtx, isLazyComputed = false) {
-  return (computedItem, computedHandler, depKeys, compare, sort) => {
-    // if user don't pass sort explicitly, computed fn will been called orderly by sortFactor
-    // sort param may in computedHandler when it is an object like {fn:()=>{}, sort:10}
-    const _sort = (sort || sortFactor++);
-
+export default function (refCtx) {
+  return (computedItem, computedHandler, depKeysOrOpt) => {
     const confMeta = {
-      type: 'computed', isLazyComputed, refCtx, sort: _sort,
+      type: FN_CU, refCtx,
       stateKeys: refCtx.stateKeys, retKeyFns: refCtx.computedRetKeyFns,
       module: refCtx.module, connect: refCtx.connect, dep: refCtx.computedDep
     };
-    configureDepFns(CATE_REF, confMeta, computedItem, computedHandler, depKeys, compare);
+    configureDepFns(CATE_REF, confMeta, computedItem, computedHandler, depKeysOrOpt);
   };
 }