+++ src/core/watch/get-define-watch-handler.js
@@ -1,12 +1,12 @@
 import configureDepFns from '../base/configure-dep-fns';
-import { CATE_REF } from '../../support/constant';
+import { CATE_REF, FN_WATCH } from '../../support/constant';
 
 export default function (refCtx) {
-  return (watchItem, watchHandler, depKeys, compare, immediate) => {
+  return (watchItem, watchHandler, depKeysOrOpt) => {
     const confMeta = {
-      type: 'watch', refCtx, stateKeys: refCtx.stateKeys, retKeyFns: refCtx.watchRetKeyFns,
+      type: FN_WATCH, refCtx, stateKeys: refCtx.stateKeys, retKeyFns: refCtx.watchRetKeyFns,
       module: refCtx.module, connect: refCtx.connect, dep: refCtx.watchDep
     };
-    configureDepFns(CATE_REF, confMeta, watchItem, watchHandler, depKeys, compare, immediate);
+    configureDepFns(CATE_REF, confMeta, watchItem, watchHandler, depKeysOrOpt);
   };
 }