+++ src/support/priv-constant.js
@@ -5,6 +5,8 @@ export const REF_KEY = Symbol('refk');
 export const START = '1';
 export const END = '2';
 
+export const FUNCTION = 'function';
+
 export const NOT_A_JSON = 'is not a plain json object!';
 
 export const STR_ARR_OR_STAR = 'should be an string array or *!';